requirements.txt
streamlit==0.78.0
pandas==1.2.3
numpy==1.20.1
pycaret==2.0.0
Pillow==8.0.1